split productGrid into grid and card components and added prompt telling product is added
