+++ src/components/ProductCard/ProductCard.module.css
@@ -1,0 +1,64 @@
+.card_container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  width: 200px;
+  height: 400px;
+}
+
+.image_wrapper {
+  border-radius: 20px;
+  height: 100%;
+  overflow: hidden;
+  box-shadow: 0px 5px 10px 0px black;
+}
+
+.card_image {
+  object-fit: cover;
+  object-position: center;
+  width: 100%;
+  height: 100%;
+  border-radius: 20px;
+  transition: transform 0.5s ease;
+}
+
+.card_container:hover .card_image {
+  transform: scale(1.1);
+}
+
+.card_title {
+  margin: 0;
+  margin-top: 10px;
+}
+
+.card_category {
+  margin: 0;
+  opacity: 70%;
+  font-size: 1.1rem;
+}
+
+.card_price_container {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.previousPrice {
+  margin: 0;
+  margin-bottom: 10px;
+  font-size: 1.3rem;
+  text-decoration: line-through;
+}
+
+.card_discounted_price {
+  margin: 0;
+  margin-bottom: 10px;
+  margin-left: 5px;
+  font-size: 1.5rem;
+}
+
+@media screen and (max-width: 900px) {
+  .card_grid {
+    padding-top: 20px;
+  }
+}
