+++ src/components/productGrid/index.tsx
@@ -1,16 +1,13 @@
-import { Link } from "react-router-dom";
-import ReviewStars from "../shared/reviewStars";
 import styles from "./productCard.module.css";
-import Button from "../shared/button";
 import React, { useContext } from "react";
-import { APIResult, CartContext } from "../../App";
+import { APIResult } from "../../App";
 import Loader from "../shared/loader";
+import ProductCard from "../ProductCard";
 
 function ProductGrid() {
-  const { state, dispatch } = useContext(CartContext);
   const { allProducts, loading, error } = useContext(APIResult);
 
-  if (loading && !state) {
+  if (loading) {
     return (
       <Loader/>
     );
@@ -24,51 +21,12 @@ function ProductGrid() {
     );
   }
 
-  type AddToCart = {
-    id: string,
-    title: string,
-    discountedPrice: number,
-    price: number,
-    image: {
-        alt: string,
-        url: string,
-    },
-  }
-
-  const handleAddToCart = ({id, title, discountedPrice, price, image}: AddToCart) => {
-    dispatch({type: "addToCart", payload: { id, title, discountedPrice, price, image, quantity: 1 }});
-  }
-
   return (
     <div className={styles.card_grid}>
       {allProducts.map(
-        ({ id, title, image, discountedPrice, price, rating, tags }) => {
+        (product) => {
           return (
-            <div key={id}>
-              <Link to={`/product/${id}`} className={styles.card_container}>
-                <div className={styles.image_wrapper}>
-                  <img
-                    className={styles.card_image}
-                    src={image.url}
-                    alt={image.alt}
-                  />
-                </div>
-                <h2 className={styles.card_title}>{title}</h2>
-                <p className={styles.card_category}>{tags[0]}</p>
-                <ReviewStars stars={rating} />
-                <div className={styles.card_price_container}>
-                  {price !== discountedPrice ? <p className={styles.previousPrice}>kr{price}</p> : <p className={styles.card_discounted_price}>kr{price}</p>}
-                  {price !== discountedPrice ? <p className={styles.card_discounted_price}>kr{discountedPrice}</p> : ""}
-                </div>
-              </Link>
-              <Button
-                text="Add to Cart"
-                type="button"
-                handleEvent={() => {
-                  handleAddToCart({id, title, discountedPrice, price, image});
-                }}
-              />
-            </div>
+            <ProductCard key={product.id} {...product} />
           );
         }
       )}
