+++ src/pages/cart.module.css
@@ -26,77 +26,6 @@
   margin: 0;
 }
 
-.productContainer {
-  display: flex;
-  padding-block: 20px;
-  border-bottom: 2px solid var(--tertiary);
-}
-
-.productImage {
-  min-width: 150px;
-  max-height: 150px;
-  object-fit: cover;
-  object-position: center;
-  border-radius: 20px;
-  box-shadow: 0px 2px 10px 0px black;
-}
-
-.productInfoContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.productTitleContainer {
-  display: flex;
-  justify-content: space-between;
-  margin-left: 20px;
-}
-
-.productTitle {
-  flex: 2;
-  margin: 0;
-}
-
-.productInteraction {
-  display: flex;
-  gap: 20px;
-  margin-left: 20px;
-  max-width: 400px;
-}
-
-.quantityInput {
-  max-width: 100px;
-  height: 28px;
-  background-color: var(--text);
-  border: none;
-  padding-left: 5px;
-  font-weight: bold;
-  text-align: center;
-}
-
-.priceContainer {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  gap: 5px;
-  width: 100%;
-  text-align: end;
-}
-
-.price {
-  width: 100%;
-  margin: 0;
-  font-size: 1.3rem;
-}
-
-.discountedPrice {
-  margin: 0;
-  text-decoration: line-through;
-  font-size: 1.2rem;
-}
-
 .summaryContainer {
   position: fixed;
   width: 25%;
@@ -144,46 +73,6 @@
 
   .cartContentContainer {
     width: 100%;
-  }
-
-  .productContainer {
-    text-wrap: nowrap;
-    border-bottom: 0px;
-    padding-inline: 20px;
-  }
-
-  .productImage {
-    min-width: 100px;
-    max-height: 100px;
-  }
-
-  .productInfoContainer {
-    padding-left: 0;
-  }
-
-  .productTitle {
-    font-size: 1.1rem;
-    text-wrap: wrap;
-  }
-
-  .productInteraction {
-    display: none;
-  }
-
-  .productInteractionMobile {
-    display: flex;
-    gap: 10px;
-    padding-inline: 20px;
-    padding-bottom: 10px;
-    border-bottom: 2px solid var(--tertiary)
-  }
-
-  .price {
-    font-size: 1.1rem;
-  }
-
-  .discountedPrice {
-    font-size: 1rem;
   }
 
   .summaryContainer {
